+++ vaadin-dialog-flow-integration-tests/src/main/java/com/vaadin/flow/component/dialog/tests/CloseListenerReopenDialogPage.java
