+++ vaadin-dialog-flow/src/test/java/com/vaadin/flow/component/dialog/DialogSerializableTest.java
@@ -1,4 +1,4 @@
-package com.vaadin.flow.component.dialog.tests;
+package com.vaadin.flow.component.dialog;
 
 import com.vaadin.flow.testutil.ClassesSerializableTest;
